update in preparation for timeout support
for now, this just splits the current call to RfcCallReceive into two
calls to RfcCall and RfcReceive respectively. The second one continues
to be blocking, so there's no functional change.

Do note the sneaky parameter naming issue in the current code where
what's in Importing has to go into the 'exporting' parameter of
RfcCalReceive (and the reverse, of course)